server/nordhapp/serializers.py
from rest_framework import serializers
from nordhapp.models import Products


class ProductsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'